Mejoras en el log del sever para todas las acciones y las funciones se pasaron de la vista al modelo
