+++ CRUD.log
@@ -1,3 +1,19 @@
 2024-05-26 20:36:56.951514: Alta Decorador de: AAAAAAAAA, con DNI: 22384765 y tiempo de: 06:23
 2024-05-26 20:40:01.508368: Alta Decorador de: qwd, con DNI: 21434324 y tiempo de: 33:33
 2024-05-26 20:45:52.249783: Alta Decorador de: qwd, con DNI: 75983222 y tiempo de: 32:00
+2024-05-26 20:51:33.657360: Alta Decorador de: qwd, con DNI: 88765253 y tiempo de: 12:12
+2024-05-26 20:52:11.209519: Alta Decorador de: qwd qdw, con DNI: 21384763 y tiempo de: 12:23
+2024-05-26 21:05:27.701155: Alta Decorador de: wqwd qwd, con DNI: 938372983 y tiempo de: 12:20
+2024-05-26 21:05:37.139162: Alta Decorador de: wqwd qwd, con DNI: 98876384 y tiempo de: 12:20
+2024-05-26 21:12:04.529017: Baja Decorador de: qwd, con DNI: 21434324 y tiempo de: 33:33
+2024-05-26 21:13:12.648036: Baja Decorador de: AAAAAAAAA, con DNI: 22384765 y tiempo de: 06:23
+2024-05-26 21:13:59.760802: Modificacion Decorador de: qwdwqd, con DNI: 75983222 y nuevo tiempo: 32:00
+2024-05-26 21:14:11.018790: Modificacion Decorador de: qwdwqdrre, con DNI: 75983222 y nuevo tiempo: 32:00
+2024-05-26 21:14:16.469968: Baja Decorador de: qwd qdw, con DNI: 21384763 y tiempo de: 12:23
+2024-05-26 21:18:20.369735: Alta Decorador de: qwd, con DNI: 3388772 y tiempo de: 12:12
+2024-05-26 21:18:25.389139: Modificacion Decorador de: qwdwqdrreqwd, con DNI: 75983222 y nuevo tiempo: 32:00
+2024-05-26 21:18:30.871009: Baja Decorador de: qwe, con DNI: 31223444 y tiempo de: 12:20
+2024-05-26 21:19:53.160995: Baja Decorador de: qwd, con DNI: 88765253 y tiempo de: 12:12
+2024-05-26 21:20:09.850431: Baja Decorador de: qwd, con DNI: 34223123 y tiempo de: 12:22
+2024-05-26 21:20:30.448845: Alta Decorador de: qwed wdwq, con DNI: 12998323 y tiempo de: 12:12
+2024-05-26 21:20:37.488717: Baja Decorador de: qwdwqdrreqwd, con DNI: 75983222 y tiempo de: 32:00
