+++ cliente.py
@@ -1,18 +1,19 @@
 import socket
 
-def send_log_message(message):
+def enviar_mensaje_log(message):
     HOST, PORT = "localhost", 9999
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
     sock.sendto(message.encode('utf-8'), (HOST, PORT))
 
-# Example usage in a CRUD application
 def alta_server_log(dni, nombre, tiempo_50_mts):
-    # Perform the creation of a new record in the database
-    # ...
+    loguear_mensaje = f"Alta con Server prendido: DNI={dni}, Name={nombre}, Time={tiempo_50_mts}"
+    enviar_mensaje_log(loguear_mensaje)
 
-    # Send log message to the server
-    log_message = f"Alta con Server prendido: DNI={dni}, Name={nombre}, Time={tiempo_50_mts}"
-    send_log_message(log_message)
+def modificar_server_log(dni, nombre, tiempo_50_mts):
+    loguear_mensaje = f"Modificar con Server prendido: DNI={dni}, Name={nombre}, New Time={tiempo_50_mts}"
+    enviar_mensaje_log(loguear_mensaje)
 
-# Simulate creating a record
+def borrar_server_log(dni, nombre):
+    loguear_mensaje = f"Borrar con Server prendido: DNI={dni}, Name={nombre}"
+    enviar_mensaje_log(loguear_mensaje)
